services/edge_node/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
psutil==5.9.7
paho-mqtt==1.6.1
httpx==0.26.0
python-dotenv==1.0.0
